requirements.txt
acme==0.35.0
asn1crypto==0.24.0
attrs==19.1.0
backcall==0.1.0
bcrypt==3.1.4
beautifulsoup4==4.6.0
bitarray==0.8.1
bleach==2.1.3
bootstrap4==0.1.0
botocore==1.12.167
Brotli==1.0.4
certbot==0.35.0
certifi==2018.4.16
cffi==1.11.4
chardet==3.0.4
click==6.7
cloudpickle==0.5.2
colorama==0.3.9
ConfigArgParse==0.14.0
configobj==5.0.6
contextlib2==0.5.5
cryptography==2.3.1
cycler==0.10.0
Cython==0.27.3
cytoolz==0.9.0
dask==0.16.1
decorator==4.2.1
defusedxml==0.5.0
distributed==1.20.2
dj-database-url==0.5.0
Django==2.1.7
django-bootstrap==0.2.4
django-bootstrap4==0.0.6
django-braces==1.13.0
django-crispy-forms==1.7.2
django-debug-toolbar==1.10.1
django-decouple==2.1
django-extensions==2.1.0
django-inbound-email==0.13
django-json-response==1.1.5
django-localflavor==2.1
django-storages==1.7.1
django-urls==1.0.6
django-widget-tweaks==1.4.2
docutils==0.14
entrypoints==0.2.3
et-xmlfile==1.0.1
fabric==2.4.0
fastcache==1.0.2
filelock==2.0.13
future==0.17.1
gevent==1.2.2
glob2==0.6
greenlet==0.4.12
gunicorn==19.9.0
HeapDict==1.0.0
html5lib==1.0.1
idna==2.7
invoke==1.1.1
ipykernel==5.1.1
ipython==7.5.0
ipython-genutils==0.2.0
itsdangerous==0.24
jedi==0.13.3
Jinja2==2.10.1
jmespath==0.9.3
josepy==1.1.0
jsonschema==3.0.1
jtutils==0.0.6
jupyter-client==5.2.4
jupyter-core==4.4.0
kiwisolver==1.1.0
leven==1.0.4
Logbook==1.4.0
lxml==4.2.5
Markdown==3.0.1
MarkupSafe==1.1.1
mistune==0.8.4
mock==3.0.5
msgpack-python==0.5.1
nbconvert==5.5.0
nbformat==4.4.0
ngrok==0.0.1
nose==1.3.7
notebook==5.7.8
numpy==1.16.3
pandocfilters==1.4.2
paramiko==2.4.2
parsedatetime==2.4
parso==0.4.0
pbr==4.2.0
pexpect==4.7.0
pickleshare==0.7.5
Pillow==5.2.0
pipenv==2018.11.26
pluggy==0.7.1
prettytable==0.7.2
prometheus-client==0.6.0
prompt-toolkit==2.0.9
psutil==5.4.3
psycopg2==2.7.5
psycopg2-binary==2.7.5
ptyprocess==0.6.0
py==1.6.0
pyasn1==0.4.4
pycparser==2.18
Pygments==2.4.0
PyNaCl==1.2.1
pyOpenSSL==18.0.0
pyparsing==2.4.0
pyRFC3339==1.1
pyrsistent==0.15.2
python-dateutil==2.6.1
python-decouple==3.1
python-http-client==3.1.0
python-magic==0.4.15
pytz==2018.5
PyYAML==3.13
pyzmq==18.0.1
qtconsole==4.4.1
requests==2.22.0
requests-toolbelt==0.9.1
resolver==0.2.1
rsa==3.4.2
s3cmd==2.0.2
s3transfer==0.1.13
Send2Trash==1.5.0
simplegeneric==0.8.1
six==1.11.0
sortedcontainers==1.5.9
sqlparse==0.2.4
stevedore==1.29.0
stripe==2.28.1
tblib==1.3.2
terminado==0.8.2
testpath==0.4.2
toolz==0.9.0
tornado==4.5.3
tox==3.2.1
traitlets==4.3.2
Unipath==1.1
urllib3==1.25.2
virtualenv==16.6.0
virtualenv-clone==0.5.3
virtualenvwrapper==4.8.2
wcwidth==0.1.7
webencodings==0.5.1
whitenoise==4.0
widgetsnbextension==3.4.2
Wrapper==1.1.0b1
wrapt==1.11.1
xlrd==1.2.0
xmltodict==0.12.0
zict==0.1.3
zope.component==4.5
zope.deferredimport==4.3
zope.deprecation==4.4.0
zope.event==4.4
zope.hookable==4.2.0
zope.interface==4.6.0
zope.proxy==4.3.1
